fixed after a code review
