+++ README.md
@@ -1,6 +1,6 @@
 # Movies list - Filter
 - Replace `<your_account>` with your Github username in the
- [DEMO LINK](https://<your_account>.github.io/react_movies-list-filter/)
+ [DEMO LINK](https://Layrovell.github.io/react_movies-list-filter/)
 - Follow the [React task guideline](https://github.com/mate-academy/react_task-guideline#react-tasks-guideline)
 
 ## Task
